\name{methylMnM-package}
\alias{methylMnM-package}
\alias{methylMnM}
\docType{package}
\title{
MeDIP-Seq and MRE-Seq data analysis
}
\description{
M&M was developed for analyzing data derived from methylated DNA immunoprecipitation (MeDIP) experiments followed by sequencing (MeDIP-Seq) and
the digestions with the methyl-sensitive restriction enzymes (MRE-Seq). 
	Nevertheless, functionalities like the quality controls may be applied to other types of sequencing data (e.g. ChIP-Seq). 
	MeDIP-MRE (methylMnM) test which combine the two differential techniques
(MeDIP-seq and MRE-seq) data to detecting the differentially methylation level of CpG.
}
\details{
\tabular{ll}{
Package: \tab methylMnM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-01\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.12.0)\cr
}
}
\author{
Yan Zhou, Bo Zhang, Nan Lin, BaoXue Zhang and Ting Wang

Maintainer: Yan Zhou <zhouy1016@163.com>
}
\references{
Yan Zhou, Bo Zhang, Nan Lin, BaoXue Zhang and Ting Wang, 2012
}
\keyword{ package }
