\name{qvalue.rank}
\alias{qvalue.rank}
\title{
Rank values.
}
\description{
The function is used to rank values.
}
\usage{
qvalue.rank(x)
}
\arguments{
  \item{x}{
Value.
}
}
\value{
Ranked values.
}

\author{
Yan Zhou, Bo Zhang, Nan Lin, BaoXue Zhang and Ting Wang
}
\examples{
x<-c(4,2,50,42,80,9)
qvalue.rank(x)
}

